RESTFUL API Endpoint
I have created a new API endpoint with a simple controller that handles the extraction of text when the PDF is uploaded.

I have also split up the functions from the controller into a helper and a service for mockery later when testing.

There’s a migration in this for the model of payslip extractions as I’ll be using this as my example.